+++ src/frontend/components/app/sidebar/sidebar-parent.tsx
@@ -2,8 +2,9 @@ import React from 'react'
 
 import SidebarResource from './sidebar-resource'
 import ResourceJSON from '../../../../backend/decorators/resource-json.interface'
-import { NavGroup } from '../../design-system'
+import { NavGroup, Box } from '../../design-system'
 import { useTranslation } from '../../../hooks'
+import { cssClass } from '../../design-system/utils/css-class'
 
 
 type Props = {
@@ -18,6 +19,19 @@ const SidebarParent: React.FC<Props> = (props) => {
   const { parent } = props
   const { icon, name, resources } = parent
   const { translateLabel } = useTranslation()
+
+  if (!parent.name) {
+    return (
+      <Box pl="default" pb="xl" pt="sm" ml="sm" className={cssClass('SidebarParent')}>
+        {resources.map(resource => (
+          <SidebarResource
+            resource={resource}
+            key={resource.id}
+          />
+        ))}
+      </Box>
+    )
+  }
 
   return (
     <NavGroup icon={icon} title={translateLabel(name)}>